+++ ReportingTests/DataMunging/Reports/ReportMinMaxTemperature.cs
@@ -7,11 +7,11 @@ using DataMunging.Reporting.ViewModels;
 namespace DataMunging.Reporting.Reports
 {
     // Given a collection of DayDayId, Min, and Max values, report min and max spreads
-    public class ReportMinMaxTemperature<T> where T : IEnumerable
+    public class ReportMinMaxTemperature<T> where T : IDailyTemperatures
     {
-        private readonly T points;
+        private readonly IEnumerable<T> points;
 
-        public ReportMinMaxTemperature(T points)
+        public ReportMinMaxTemperature(IEnumerable<T> points)
         {
             this.points = points;
         }