+++ FlatFile.Core/ChangeLog.md
@@ -681,7 +681,7 @@ var layout = new LayoutDescriptor<LeagueScore>()
 ## 11/12/18
 - Mocked expected result for league standings, and tested `LeagueStandingsImporter` against `football.dat`
 - Updated `ReportMinMax` to work with `Type T` vs. just `IEnumerable<IDailyTemperatures>`
+- Added report for scoring spreads with `ReportMinMaxScore`
 
-- resume above layout desciptor and test import. Then run report. 
+Resume combining `ReportMinMaxScore` and `ReportMinMaxTemperature`
 
-Then, pt 2 of data munging